+++ maven-core/src/test/projects/lifecycle-executor/project-with-multiple-executions/src/main/mdo/remote-resources.mdo
@@ -20,10 +20,10 @@ under the License.
 <model>
   <id>remoteResourcesBundle</id>
   <name>RemoteResourcesBundle</name>
-  <description><![CDATA[
+  <description>
     A bundle of remote resources.
     This is the model specification for remote-resources.xml.
-  ]]></description>
+  </description>
   <defaults>
     <default>
       <key>package</key>
@@ -39,9 +39,9 @@ under the License.
         <field>
           <name>remoteResources</name>
           <version>1.1.0</version>
-          <description><![CDATA[
+          <description>
             Remote resources that are contained in a remote resources bundle.
-          ]]></description>
+          </description>
           <association>
             <type>String</type>
             <multiplicity>*</multiplicity>
@@ -50,9 +50,9 @@ under the License.
         <field>
           <name>sourceEncoding</name>
           <version>1.1.0</version>
-          <description><![CDATA[
+          <description>
             Source encoding of the remote resources contained in the bundle.
-          ]]></description>
+          </description>
           <type>String</type>
         </field>
       </fields>
